long long types fixed in exit status

+++ LIBFT/libft.h
@@ -46,7 +46,7 @@ char				*ft_strrchr(const char *str, int c);
 char				*ft_strnstr(const char *str,
 						const char *needle, size_t len);
 int					ft_strncmp(const char *str1, const char *str2, size_t n);
-long long			ft_atoi(const char *str);
+int					ft_atoi(const char *str);
 int					ft_isalpha(int c);
 int					ft_isdigit(int c);
 int					ft_isalnum(int c);
